Refactor game script types into generic engine.d.ts
Split Breakout-specific types into breakout.d.ts and add getScreen/isActiveScreen
helpers so multi-screen state access is typed without confusing state.screen
(active screen name) with the injected screen global (framebuffer size).

Engine types (GameState, MultiScreenState, GameScript, SpriteDef, etc.) now
live in engine.d.ts; game.d.ts is a thin entry reference. Breakout example
updated to use TypedEventProps and game_helpers.
