+++ gallery/game_engine/scripting/game_helpers.tsx
@@ -1,0 +1,28 @@
+/// <reference path="./engine.d.ts" />
+//
+// Small runtime helpers for multi-screen game scripts.
+// Import from game scripts: import { getScreen, activeScreen } from "./game_helpers";
+//
+
+/** Active screen name (`state.screen`). */
+export function activeScreen<TActive extends string>(
+  state: { screen: TActive }
+): TActive {
+  return state.screen;
+}
+
+/** Typed access to a named screen's frozen sub-state (`state.screens[name]`). */
+export function getScreen<
+  TScreens extends object,
+  K extends keyof TScreens & string
+>(state: { screens: TScreens }, name: K): TScreens[K] {
+  return state.screens[name];
+}
+
+/** True when `state.screen` equals `name` (narrows in TypeScript). */
+export function isActiveScreen<TActive extends string>(
+  state: { screen: string },
+  name: TActive
+): state is { screen: TActive } {
+  return state.screen === name;
+}
